# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "NodeJS API",
    "version": "1.0.0",
    "description": "API that is used to perform CRUD operations for a todolist application in the OpenAPI 3.0 specification.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Tanel Märjama",
      "email": "tanel.marjama@voco.ee"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/login": {
      "summary": "Log user in.",
      "description": "Endpoint to log user in and give authorization to edit the todo items.",
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userLoggedInResponse"
                }
              }
            },
            "description": "User logged in."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            },
            "description": "No user found with those credentials."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "operationId": "authorizeUser",
        "summary": "Log user in.",
        "description": "Endpoint to log a user in."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "login": {
        "title": "Root Type for login",
        "description": "Data to log in a user.",
        "type": "object",
        "properties": {
          "username": {
            "description": "Users username.",
            "type": "string"
          },
          "password": {
            "description": "Users password.",
            "type": "string"
          }
        },
        "example": {
          "username": "User1",
          "password": "Password1"
        }
      },
...
      "userLoggedInResponse": {
        "title": "Root Type for userLoggedIn",
        "description": "Data returned from the server if a user is logged in.",
        "type": "object",
        "properties": {
          "id": {
            "format": "int32",
            "description": "ID of the user.",
            "type": "integer"
          },
          "token": {
            "description": "Token for user authentication",
            "type": "string"
          }
        },
        "example": {
          "id": 1,
          "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJVc2VySUQiOjIsImlhdCI6MTY3NDA0NTA2M30.uNP-MYOO2r0pnPBsblJxkZ0Eu85I532oYQR66wMWE2s"
        }
      },
      "error401": {
        "description": "Errormessage for unauthorized status.",
        "type": "object",
        "example": "\"Unauthorized. Please try logging in again.\""
      },
      "error500": {
        "description": "Errormessage for internal server error.",
        "type": "object",
        "example": "\"Internal server error.\""
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "type": "http",
        "description": "Authorization for API"
      }
    }
  }
}